feat(audio) Uses memory pool for audio data to reduce allocations
This increases the throughput of the audio capture stage. Because of
this, I've also had to change the audio encoding to a timer based approach
instead of using notifications from the audio capture thread. This is
because Pipewire can sometimes deliver small chunks of audio very quickly
and this was saturating the processing loop, causing high CPU usage.