feat(metrics) Adds timing metrics collection for audio & video
Also adds some performance testing results, incorporating the
metrics.

This must be enabled at build time using
`-DSHADOW_CAST_ENABLE_METRICS=ON`. Metrics are written to
a file in the same directory as the output media, using the same
filename, but with a `.metrics` suffix.

Also added a simple `gnuplot` script (`./tools/metrics-plot`) for
visualizing the data.