feat: (Exp.) Use a LRU cache for image buffers
Now we have explicit sync, we can re-use DRM buffers instead of
recreating them every frame. We now keep a small LRU eviction cache
of DRM buffers. This buffer will be flushed of stale DRM buffers every
so often so as not to keep them from consuming resources. The size
maximum size of the cache can be controlled via the `-C` cmdline param.